Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
torch
numpy
pandas
